+++ src/pages/works/import/WorksImportPage.vue
@@ -17,18 +17,6 @@ const importMethods = [
     sectionTitle: 'Import from NaNoWriMo',
     methods: [
       {
-        title: 'Automatic NaNoWriMo Import',
-        icon: PrimeIcons.SYNC,
-        details: [`Import all your NaNoWriMo projects at once by letting TrackBear log in on your behalf.`],
-        route: 'import-works-nano-auto',
-      },
-      {
-        title: 'Automatic Young Writers Program Import',
-        icon: PrimeIcons.SYNC,
-        details: [`Import all your NaNoWriMo Young Writers Program projects at once by letting TrackBear log in on your behalf.`],
-        route: 'import-works-ywp-auto',
-      },
-      {
         title: 'Manual Import',
         icon: PrimeIcons.COPY,
         details: [`Import a single project by copying and pasting data from NaNoWriMo's website.`],
@@ -36,23 +24,25 @@ const importMethods = [
       },
     ],
   },
-  // {
-  //   sectionTitle: 'Import from elsewhere',
-  //   methods: [
-  //     {
-  //       title: 'Manual Import',
-  //       icon: PrimeIcons.COPY,
-  //       details: [`Import a single project by copying and pasting data into TrackBear.`],
-  //       route: 'import-works-manual',
-  //     },
-  //     {
-  //       title: 'Upload a CSV',
-  //       icon: PrimeIcons.UPLOAD,
-  //       details: [`Import a single project by uploading a CSV file.`],
-  //       route: 'import-works-csv',
-  //     },
-  //   ],
-  // },
+  {
+    sectionTitle: 'Import from anywhere (Coming soon!)',
+    methods: [
+      {
+        title: 'Upload a CSV',
+        icon: PrimeIcons.UPLOAD,
+        details: [`Import a single project by uploading a CSV file.`, `(Coming soon!)`],
+        // route: 'import-works-csv',
+        route: 'import-works',
+      },
+      {
+        title: 'Manual Import',
+        icon: PrimeIcons.COPY,
+        details: [`Import a single project by copying and pasting data into TrackBear.`, `(Coming soon!)`],
+        // route: 'import-works-manual',
+        route: 'import-works',
+      },
+    ],
+  },
 ];
 
 </script>
@@ -74,7 +64,7 @@ const importMethods = [
           :to="{ name: method.route }"
         >
           <Card
-            class="max-w-64"
+            class="max-w-half"
             :pt="{ content: { class: '!py-0' } }"
             :pt-options="{ mergeSections: true, mergeProps: true }"
           >